adding vnic rx tx

+++ kernel/dev/x86-64/network/e1000/e1000.c
@@ -29,13 +29,13 @@ uint8_t e1000_init(struct device* dev) {
     return 1;
 }
 
-void e1000_ethernet_read(struct device* dev, uint8_t* data, uint16_t size) {
+void e1000_ethernet_read(struct device* dev, uint64_t* data, uint16_t size) {
     ASSERT_NOT_NULL(dev);
     ASSERT_NOT_NULL(data);
 
     panic("Ethernet read not implemented yet");
 }
-void e1000_ethernet_write(struct device* dev, uint8_t* data, uint16_t size) {
+void e1000_ethernet_write(struct device* dev, uint64_t* data, uint16_t size) {
     ASSERT_NOT_NULL(dev);
     ASSERT_NOT_NULL(data);
