+++ kernel/dev/x86-64/network/ne2000pci/ne2000pci.c
@@ -153,13 +153,13 @@ uint8_t ne2000_pci_init(struct device* dev) {
     return 1;
 }
 
-void ne2000pci_ethernet_read(struct device* dev, uint8_t* data, uint16_t size) {
+void ne2000pci_ethernet_read(struct device* dev, uint64_t* data, uint16_t size) {
     ASSERT_NOT_NULL(dev);
     ASSERT_NOT_NULL(data);
 
     panic("Ethernet read not implemented yet");
 }
-void ne2000pci_ethernet_write(struct device* dev, uint8_t* data, uint16_t size) {
+void ne2000pci_ethernet_write(struct device* dev, uint64_t* data, uint16_t size) {
     ASSERT_NOT_NULL(dev);
     ASSERT_NOT_NULL(data);
